Move webauthn public model to api/model
webauthn package will depend on lib/authn/user
and lib/authn/user depends on lib/authn/identity.
To break the dependency cycle, move the public models to api/model.
