+++ pkg/lib/authn/authenticator/passkey_authenticator.go
@@ -4,19 +4,18 @@ import (
 	"time"
 
 	"github.com/authgear/authgear-server/pkg/api/model"
-	"github.com/authgear/authgear-server/pkg/lib/webauthn"
 )
 
 type Passkey struct {
-	ID                  string                    `json:"id"`
-	UserID              string                    `json:"user_id"`
-	CreatedAt           time.Time                 `json:"created_at"`
-	UpdatedAt           time.Time                 `json:"updated_at"`
-	Kind                string                    `json:"kind"`
-	IsDefault           bool                      `json:"is_default"`
-	CredentialID        string                    `json:"credential_id"`
-	CreationOptions     *webauthn.CreationOptions `json:"creation_options,omitempty"`
-	AttestationResponse []byte                    `json:"attestation_response,omitempty"`
+	ID                  string                         `json:"id"`
+	UserID              string                         `json:"user_id"`
+	CreatedAt           time.Time                      `json:"created_at"`
+	UpdatedAt           time.Time                      `json:"updated_at"`
+	Kind                string                         `json:"kind"`
+	IsDefault           bool                           `json:"is_default"`
+	CredentialID        string                         `json:"credential_id"`
+	CreationOptions     *model.WebAuthnCreationOptions `json:"creation_options,omitempty"`
+	AttestationResponse []byte                         `json:"attestation_response,omitempty"`
 	// SignCount of 0 means sign count is not supported by the authenticator.
 	// So we do not include omitempty here.
 	SignCount int64 `json:"sign_count"`
