+++ pkg/lib/authn/authenticator/passkey_spec.go
@@ -1,11 +1,11 @@
 package authenticator
 
 import (
-	"github.com/authgear/authgear-server/pkg/lib/webauthn"
+	"github.com/authgear/authgear-server/pkg/api/model"
 )
 
 type PasskeySpec struct {
-	AttestationResponse []byte                    `json:"attestation_response,omitempty"`
-	AssertionResponse   []byte                    `json:"assertion_response,omitempty"`
-	CreationOptions     *webauthn.CreationOptions `json:"creation_options,omitempty"`
+	AttestationResponse []byte                         `json:"attestation_response,omitempty"`
+	AssertionResponse   []byte                         `json:"assertion_response,omitempty"`
+	CreationOptions     *model.WebAuthnCreationOptions `json:"creation_options,omitempty"`
 }
